+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Adding a class method `configure!` to configure the matcher
+
 # 1.0.0
 
 * Log matcher added